consolidate CLI entry points
